third_party/blink/web_tests/external/wpt/html/webappapis/scripting/processing-model-2/unhandled-promise-rejections/promise-rejection-event-during-parse-expected.txt

This is a testharness.js-based test.
FAIL Promise rejection during initial parsing of document assert_array_equals: expected property 0 to be "readystatechange:interactive" but got "unhandledrejection" (expected array ["readystatechange:interactive", "unhandledrejection", "readystatechange:complete"] got ["unhandledrejection", "readystatechange:interactive", "readystatechange:complete"])
Harness: the test ran to completion.
